Move game element classes
